requirements.txt
telethon>=1.0
python-dotenv